chore: ts conversion
