+++ js/src/admin/utils/frontendLabel.ts
@@ -1,13 +1,13 @@
 import app from "flarum/admin/app";
 
-export default function (frontend) {
+export default function frontendLabel(frontend: string): string {
     const translation =
         app.translator.translations[
             "fof-linguist.admin.known-frontend." + frontend
         ];
 
     if (translation) {
-        return translation;
+        return translation as string;
     }
 
     return frontend;
